[refactor] 폼 제출 및 validateField 함수 form.js 로 이동 및 로그인도 공용 함수 활용하도록 리팩토링
